Added Branch Renaming
I've explained the difference between "git branch -m" and "git branch -M". I've also explained the reason why we rename the default "master" branch to "main". Lastly, I've added an easier way to reset to the last commit(using head~1) and to reset the GitHub repository to the specified commitID.